+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 
 ## 2.12.0 (WIP)
 
-* `Misc` Add-on is now using Manifest V3.
+* `Feature` Add-on is now using Manifest V3.
+* `Fix` Context menu to delete or move emote did not show in emote box.
 * `Misc` Removed more permissions.
 * `Misc` Cleaned up code for CSS updater.